complete builder.py refactor: remove references to builder2, midas2
