+++ python/nistoar/pdr/preserv/bagger/__init__.py
@@ -8,5 +8,5 @@ create bags (via the BagBuilder class).  For example, the MIDASBagger
 understands how to bag up data provided by MIDAS.  
 """
 
-from midas2 import MIDASMetadataBagger
+from midas import MIDASMetadataBagger
 from prepupd import UpdatePrepService
