+++ python/tests/nistoar/pdr/preserv/bagger/test_midas_update.py
@@ -16,7 +16,7 @@ from nistoar.testing import *
 from nistoar.pdr import utils
 import nistoar.pdr.preserv.bagit.builder as bldr
 from nistoar.pdr.preserv.bagit.bag import NISTBag
-import nistoar.pdr.preserv.bagger.midas2 as midas
+import nistoar.pdr.preserv.bagger.midas as midas
 import nistoar.pdr.exceptions as exceptions
 from nistoar.pdr.preserv import AIPValidationError
